Initial Project Commit
Initial commit of project. Folder includes no XMLs pre-added for use, so they will need to be generated by nmap and saved in the folder in which the .py code is contained to be accessed.

From there, the code will parse the XML file(s) and sort it based on the contents of the other XMLs used, showing inconsistencies between different access types of a given network (wifi, ethernet, external).